State of game at GMTK Submission
This is the state of the game at the deadline for the GMTK submission. The game is incomplete and is missing a final ending page telling the player if they have won or lost. Additionally, I did not have time to add the sound effects (only added the music tracks).

We intend to finish the game during free time.